Image submit button's notice technique for correct alt technique should be G94, not G82.
Technique G82 ("Providing a text alternative that identifies the purpose of the non-text content") states the text alternative should be useful "even if the full function of the non-text content cannot be provided". This suggests that flagging G82 should be done for user controls where functionality is lost (like an applet).
Technique G94 (the main general technique re: short text alternative to non-text content) is probably more appropriate for an image submit button, as no functionality is lost - with alt text, it should still act as a submit button. That it should state the purpose and not the visual content is still consistent with this technique (see technique's first example: a magnifying glass icon for a search saying "Search").