Add doc entries for the following techniques in Success Criterion 1.1.1: H36 (TODO), H37 (not examples yet), H67, G94.
H36 is for adding alt text to image submit buttons.
H37 is for adding alt text to img elements; H67 is for ensuring that decorative images have empty alt text and title attribute.
G94 is a general technique thrown as a reminder that the text alternative must convey the same meaning and purpose as the non-text content.