Added test for SC 1.3.1 regarding scope attributes n tables (H63).
It is assumed that all tables found are data tables.
TH elements are flagged with an error if a scope attr is missing, or not correct (one of row, col, rowgroup, colgroup).
TD elements with the scope attr are flagged with a warning about using scope in the first place (or rather, why you aren't using TH instead) - this is due to scope being obsolete on TD elements in HTML5.
However, TD elements including scope are then checked for a valid value, because scope on td is still legal in HTML4.