HTMLCS now accepts a "_top" token which matches the top element.
The top element will now match a "_top" token which is registered like a tag name in a sniff. It will only match "_top" and not its tag name (if any), because it is technically not part of the document being tested. However, it gives the ability to test for elements anywhere in the document if need be (as is required for some tests).
The top element is the document object for the full doc string, or the enclosing div tag for a doc fragment string, or the passed DOMNode if one is passed instead.