New notice for tabindexes as per technique H4.
Checkpoint 2.4.3 ("Focus Order", single-A level) has a technique H4, "Creating a logical tab order through links, form controls, and objects". The test for this is "If tabindex is used, check that the tab order specified by the tabindex attributes follows relationships in the content." This can't be automatically checked, but a notice will be thrown if tabindex is used in the document fragment.
We use the Selector API (.querySelector method) to find an element containing the tabindex attr - if there is one, we trigger this notice.