Test for links that open in a new window (H83.3).
Starts a new sniff for Success Criterion 3.2.5 ("Change on Request"), a Triple-A criterion that basically states "Don't change context automatically unless there is the ability to turn it off" - ie. an option must exist for all changes in context to be initiated by user request only.

Technique H83 suggests use of the target="_blank" attribute - deprecated in HTML4/XHTML1.0 Transitional - removed in Strict DTDs of these due to non-support of framesets - but reinstated in HTML5 for limited use (eg. "_blank" for new windows, "_parent" in iframes) and states that the user should be warned in link text that the link opens in a new window.

This sniff provides a warning if the target is _blank and the words "new window" don't appear. It is only a warning due to the possibility of it being phrased differently, but this is the most common method of putting it.