New sniff for SC 2.4.8 (Location); test for link elements (H59).
Only applies at Triple A level.
Technique H59 relates to the usage of link elements for declaring navigation semantically: eg. using "rel" values of next, prev, etc.
This test checks ALL link elements, not just those relating to navigation, as new "rel" values for navigation may be created as part of the HTML5 living standard process.
This test checks for the elements stated in the technique - that link elements exist as children of the head element, and contain at least rel and href attributes.
(The structure "head > noscript > link" is also allowed as this is explicitly declared legal in HTML5.)
This is also an advisory technique for SC 2.4.5 (Multiple Ways) at Double A level.