Added technique H91 for SC 4.1.2 (Name/Role/Value).
H91 ensures that specific attributes or tags are available to determine a name and value for an accessibility API, as part of the name/role/value structure.
Note that these do not correspond to the "name" and "value" attributes on the element in many cases - for instance, an input field's "name" is its label or title attribute; a a link it is its text contents or title (and the value is its href).