Add SC 2.1.1; warnings about mouse-based events in HTML attrs.
Mouse-based events should have redundant functionality in keyboard actions, so they can be accessed without a mouse. The warnings give suggested keyboard-friendly events, but the warning will appear regardless of whether they exist - to remind the user that they should provide equivalent functionality as the mouse event (it might not).
Note that onclick is not a mouse-based event, despite the name. It maps to the default action of an element, and that can often be implemented with a key press.