Add general notices for the seizure-related success criteria: 2.3.1, 2.3.2.
Both criteria pass if nothing in your content flashes more than 3 times per second.
2.3.1 (the A version) also allows flashing if it is constrained to a sufficiently small area on the screen for it to fill no more than 25% of a 10-degree visual field, representing the central area of vision of the eye (see technique G176).
The message is deliberately vague as "sufficiently small" depends on the size and resolution of the monitor. The technique sets a value based on a baseline of 1024x768 on a 15-17 inch screen, and a formula for other screens of known resolution.
2.3.2 (the AAA version) does not have this threshold exception - it is 3 times per second, no exception.