Added SC 3.1.5; changed notation for multiple techniques.
SC 3.1.5 is about providing an easier-to-read version of the content, or supplementary information, when it requires more than about lower-secondary reading skills to understand.
Also changed the notation where multiple techniques are applicable. Previously it would have been written as, eg. "H39H73" for techniques H39 and H73 (either together or separately, just that they are both applicable). Now they are separated with commas.
This allows the technique code to be more easily parsed, when an interface wants to provide information on each technique separately.