Fix class name for 2.4.6; changed message targets for 2.4.2.
Previously all targets for 2.4.2 were on the html element; now it is on the lowest applicable level. For instance, if the complaint is the title is empty, it will fire on the title element. If there is no title element at all but there is a head, it will fire on the head element.