Fixed the handling of disabled level switches on the landing page of the Auditor.
Disabled switches could still be selected, and this resulted in multiple instances of the "disabled" class appearing on the View Report button. Only one of these was being removed each time a combination existed that could enable it.
Therefore if there was a situation where, eg. with no errors, if you turned off warnings, then errors, two disabled classes would be added. If you then enable notices, it would not remove all disabled classes until either other switch was switched back on.
Provide better feedback for disabled level switches (it now appears off and unswitchable) - which should avoid the possibility of doubled disabled classes in the first place - and also fix the regex that removes the disabled class.