Add a simple testing framework into the PhantomJS contrib directory.
See the changes to the SC 4.1.2 sample test file for an example of what
should be added as a comment to the test file to expose its name.

This script will read that file through the file system, read its test
name, the standard to use, and the assertions. Then it will run the file
through its web browser and match the messages up against the
assertions.

Assertions should be written in the following format:
  Assert: [No] Level Code [on #Id]

Where Level is either "Error", "Warning", or "Notice" (not case
sensitive).
Code is the message code with "*" serving as a wildcard (typically used
at the start of the code),
Including "on #Id" targets an element with a particular ID (if omitted,
the assertion will be triggered on any element); the "on" is not
case-sensitive,
and the "No" prefix (not case sensitive) inverts the condition - such
that if the condition is true it will fail. Normally, a condition will
fail unless it is triggered at least once.

Method of invocation:
$ cd Contrib/PhantomJS
$ phantomjs HTMLCS_Test.php [test path]
where [test path] is based on a root of the Tests/ directory.