Review issue #100.
The test that flags "for" attributes on labels that don't point to a form control is now a warning, not an error. It is not a fail in itself (especially now explicit labels are not necessary), but often suggests a misused or typoed for attribute and thus should be manually checked.

Also no error will appear if the "for" attribute is empty (for="").