WCAG2 3.2.2: Fix false positive for buttons without a type attribute.
Fixes Github issue: squizlabs/HTML_CodeSniffer#136.

The default for buttons without a type attribute (or an invalid option,
reverting to the default) is actually a submit button.

Because there's no way to querySelector() elements WITHOUT an attribute,
BUTTON elements are now compared as "number of BUTTON elements in total
vs. number that have type attribute of 'reset' or 'button'" - the other
two possible valid values.

BUTTON elements, however, won't be looked for at all if it can be
short-circuited by testing of INPUT elements, though (which has to have
a type attribute, because the default field type for INPUT is text).