First stage of improving the HTMLCS web site.
Taking advantage of Jekyll to be able to reduce most of the pages to
a Markdown [kramdown parser] used by Github Pages.

Layouts provided are "index" for the main page, and "inside" for all the
inside pages.

Proof of concept for the initial build is for the front page, and the
Patches folder.

Also removed the old pages.