Changes for the auditor's "additional information".
The additional information for the WCAG tests will now show the Success
Criterion rather than the Principle, and will link to the appropriate SC
in the main www.w3.org/TR/WCAG20 document (which will put you right next
to the supplementary "How To Meet" and "Understanding" documents).

To make room for this, instead of the additional information being a
mini-table, the name will be on a different line. (This also allows
"Success Criterion" to fit in full.)

This will resolve issue #142.