Added 3 new methods to HTMLCS.util to support presentation roles and aria-hidden attributes for #149 and #151.
isAccessibilityHidden() tests whether a given element is visible to an accessibility API.

isVisuallyHidden() was previously isHidden() and tests only if the element is visible in the browser.

getAllElements() replaces a private method in HTMLCS called _getAllTags(). The function accepts a parent element and returns an array of child elements that are not hidden from accessibility APIs. HTMLCS uses this as an initial gather of elements to test, but there are still many core query selectors used in individual sniffs which may need to be made aware of hidden elements and react accordingly.

Intially these seems to work with most common use case of avoiding tests of an element and all of it's children.